Move fields and instruction strategies managers to payment method folder
